+++ app/Http/Controllers/Danabus/SearchController.php
@@ -5,25 +5,34 @@ namespace App\Http\Controllers\Danabus;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use App\Http\Controllers\Controller;
-use App\Models\Route;
-
+use App\Models\Direction;
 
 class SearchController extends Controller
 {
-    public function index(){
-    	return view('danabus.search.index');
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        return view('danabus.search.index');
     }
 
-    public function search(Request $request){
-    	if ($request->ajax()) {
-    		$route_id = $request->get('$route_id');
-    		$status = $request->get('$status');
-    		$route = Route::with(['directions' => function ($query) use ($route_id, $status){
-            return $query->where('status','=',$status);
-        },'directions.stop'])->findOrFail($route_id);
-         return response()->json($route );
-    	}
-
-    	
+   /**
+     * Store a newly created resource in storage.
+     *
+     * @param \Illuminate\Http\Request $request of form
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function search(Request $request)
+    {
+        if ($request->ajax()) {
+            $route_id = $request->get('$route_id');
+            $status = $request->get('$status');
+            $directions = Direction::where('status', '=', $status)->where('route_id', '=', $route_id)->with('stop')->get();
+            return response()->json($directions);
+        }
     }
 }
