+++ resources/lang/en/search.php
@@ -1,0 +1,8 @@
+<?php
+return [
+    'search_busstop' => 'Tìm trạm xe ',
+    'type_your_address'=> 'Nhập địa điểm của bạn ',
+    'choice_route' => 'Chọn tuyến ',
+    'choice_status' => 'Chọn hướng đi ',
+    'search' => 'Tìm '
+    ];
