+++ routes/web.php
@@ -36,5 +36,5 @@ Route::group(['namespace' => 'Danabus'], function () {
     Route::get('/feedback', 'FeedbackController@index');
     Route::post('/feedback', 'FeedbackController@store');
     Route::get('/search', 'SearchController@index');
-    Route::post('/search', 'SearchController@search'); 
+    Route::post('/search', 'SearchController@search');
 });
